requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.21.8
botocore==1.24.8
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-dotenv==1.4.2
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
Pillow==9.0.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
